Place ShareVerifier into separate folder to avoid circular dependencies
